// MobileAPP/src/utilis/Interfaces.ts
// profile interfaces
export interface MatchesProfile {
  _id?: string;
  src?: any;
  firstName?: string;
  lastName?: string;
  age?: string;
  gender: string;
  basicDetails?: any;
  profileId?: string | number;
  uuid?: string;
  membership?: string;
  avatar?: string;
  locationDetails?: any;
  religionDetails?: any;
  careerDetails?: any;
}